VGMRgn: Be able to have more than 256 samples
This seems to be a good old-fashioned typo. Spotted when trying to get
WarioWare: Smooth Moves to work.

+++ src/main/VGMRgn.cpp
@@ -69,7 +69,7 @@ void VGMRgn::SetUnityKey(uint8_t theUnityKey) {
   unityKey = theUnityKey;
 }
 
-void VGMRgn::SetSampNum(uint8_t sampNumber) {
+void VGMRgn::SetSampNum(int sampNumber) {
   sampNum = sampNumber;
 }
